requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.2.0
xrpl-py==4.0.0
httpx==0.27.0
pyjwt==2.9.0
